justeringer fra live qa
